+++ src/components/blocks/Testimonial.tsx
@@ -8,8 +8,8 @@ type TestimonialProps = {
 
 export const Testimonial = ({ author, role, children }: TestimonialProps) => {
   return (
-    <figure className="border-l-2 pl-4 my-4">
-      <blockquote className="italic text-sky-400">{children}</blockquote>
+    <figure className="my-4 border-l-2 pl-4">
+      <blockquote className="text-sky-400 italic">{children}</blockquote>
       <figcaption className="mt-4 font-bold">
         — {author} <span>({role})</span>
       </figcaption>